api/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
deepface==0.0.79
numpy==1.24.3
Pillow==10.0.0